+++ src/main/java/com/javPOL/magazineJava/dao/SupplierDAO/SupplierDao.java
@@ -3,5 +3,5 @@ package com.javPOL.magazineJava.dao.SupplierDAO;
 import com.javPOL.magazineJava.dao.Dao;
 import com.javPOL.magazineJava.model.Supplier;
 
-public interface SupplierDao extends Dao<Supplier, Integer> {
+public interface SupplierDao extends Dao<Supplier, Long> {
 }
