+++ src/main/java/com/javPOL/magazineJava/service/CategoryServiceImpl.java
@@ -29,7 +29,7 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public Category findById(int id) {
+    public Category findById(Long id) {
         return categoryDao.findById(id);
     }
